Transliteration of lambda_calculus to Rust
Nothing done for performance as of yet (especially given that is not the
goal for this component).

I'm currently reviewing available papers / documents on the subject
(given that currently the links on the README are unavailable), and
would like to eventually structure the Term enum so it's typed, and thus
strongly normalising (which means that substitute will have to
terminate).

Currently no code for pretty printing, although that will be added soon.